Refactor entrypoint bootstrap to TypeScript app
