+++ index.html
@@ -69,6 +69,6 @@
     </script>
 
     <!-- Vite will rewrite this to /Athens/assets/*.js in production -->
-    <script type="module" src="./src/entry/landing.js"></script>
+    <script type="module" src="./src/entry/app.ts"></script>
   </body>
 </html>
